move flagd-proxy config/handler to common package
